refactor: split startup runtime service owners
